Extract paged Supabase test helper
